Minor improvements in commands and readme

+++ bootstrap-master.sh
@@ -6,16 +6,22 @@ apt-get --yes upgrade
 apt-get --yes install git screen vim telnet tcpdump python-setuptools gcc python-dev python-pip libssl-dev libffi-dev software-properties-common curl
 
 # Get ansible-2.1+, vanilla ubuntu-16.04 ansible (2.0.0.2) is broken due to https://github.com/ansible/ansible/issues/13876
-apt-add-repository -y ppa:ansible/ansible;apt-get update;apt-get install -y ansible
+ansible --version || (
+  apt-add-repository -y ppa:ansible/ansible
+  apt-get update
+  apt-get install -y ansible
+)
 
 # Kargo-cli
-git clone https://github.com/kubespray/kargo-cli.git /root/kargo-cli
-pushd /root/kargo-cli
-python setup.py install
-popd
+kargo --version || (
+  git clone https://github.com/kubespray/kargo-cli.git /root/kargo-cli
+  pushd /root/kargo-cli
+  python setup.py install
+  popd
+)
 
 # Copy/create nodes list
-cp /var/tmp/nodes ./nodes
+test -f ./nodes || cp /var/tmp/nodes ./nodes
 
 # Either pull or copy microservices repos
 cp -a /var/tmp/microservices* ./ccp/
